+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - SupplyChain finds or creates nodes based on config
 - Update graph files on disk, separate nodes and edges in dot file
 - Write graphs to webp
+- Rename SupplyChain into Scope
 
 ### Removed